Movded all the friends api logic to friends.routes.js; search is working with auth.routes.js; updated all components with these new routes
